+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * update scala-logging to [`3.9.0`](https://github.com/lightbend/scala-logging#390)
 * update jackson-databind to `2.9.5`
 * update mongodb-driver to [`3.7.0`](https://github.com/mongodb/mongo-java-driver/releases/tag/r3.7.0)
+* introduce `ToJSON.instance` to reserve `ToJSON.apply` for the typeclass instance. This change is backwards compatible only with scala >= 2.12.
 
 ## 0.9.8 (2018-05-08)